+++ private/buf/bufcli/cache.go
@@ -103,6 +103,11 @@ var (
 	//
 	// Normalized.
 	v3CacheModuleLockRelDirPath = normalpath.Join("v3", "modulelocks")
+	// v3CacheWasmRuntimeRelDirPath is the relative path to the Wasm runtime cache directory in its newest iteration.
+	// This directory is used to store the Wasm runtime cache. This is an implementation specific cache and opaque outside of the runtime.
+	//
+	// Normalized.
+	v3CacheWasmRuntimeRelDirPath = normalpath.Join("v3", "wasmruntime")
 )
 
 // NewModuleDataProvider returns a new ModuleDataProvider while creating the
@@ -133,6 +138,19 @@ func NewCommitProvider(container appext.Container) (bufmodule.CommitProvider, er
 			clientConfig,
 		),
 	)
+}
+
+// CreateWasmRuntimeCacheDir creates the cache directory for the Wasm runtime.
+//
+// This is used by the Wasm runtime to cache compiled Wasm plugins. This is an
+// implementation specific cache and opaque outside of the runtime. The runtime
+// will manage the cache versioning itself within this directory.
+func CreateWasmRuntimeCacheDir(container appext.Container) (string, error) {
+	if err := createCacheDir(container.CacheDirPath(), v3CacheWasmRuntimeRelDirPath); err != nil {
+		return "", err
+	}
+	fullCacheDirPath := normalpath.Join(container.CacheDirPath(), v3CacheWasmRuntimeRelDirPath)
+	return fullCacheDirPath, nil
 }
 
 // newWKTStore returns a new bufwktstore.Store while creating the required cache directories.