+++ private/bufpkg/bufcheck/multi_client_test.go
@@ -28,6 +28,7 @@ import (
 	"github.com/bufbuild/buf/private/pkg/slicesext"
 	"github.com/bufbuild/buf/private/pkg/stringutil"
 	"github.com/bufbuild/buf/private/pkg/tracing"
+	"github.com/bufbuild/buf/private/pkg/wasm"
 	"github.com/stretchr/testify/require"
 	"go.uber.org/zap"
 	"go.uber.org/zap/zaptest"
@@ -107,6 +108,7 @@ func testMultiClientSimple(t *testing.T, cacheRules bool) {
 	require.NoError(t, err)
 	multiClient := newMultiClient(
 		zap.NewNop(),
+		tracing.NopTracer,
 		[]*checkClientSpec{
 			newCheckClientSpec("buf-plugin-field-lower-snake-case", fieldLowerSnakeCaseClient, emptyOptions),
 			newCheckClientSpec("buf-plugin-timestamp-suffix", timestampSuffixClient, emptyOptions),
@@ -167,6 +169,7 @@ func TestMultiClientCannotHaveOverlappingRules(t *testing.T) {
 	require.NoError(t, err)
 	multiClient := newMultiClient(
 		zap.NewNop(),
+		tracing.NopTracer,
 		[]*checkClientSpec{
 			newCheckClientSpec("buf-plugin-field-lower-snake-case", fieldLowerSnakeCaseClient, emptyOptions),
 			newCheckClientSpec("buf-plugin-field-lower-snake-case", fieldLowerSnakeCaseClient, emptyOptions),
@@ -185,7 +188,7 @@ func TestMultiClientCannotHaveOverlappingRulesWithBuiltIn(t *testing.T) {
 	client, err := newClient(
 		zaptest.NewLogger(t),
 		tracing.NopTracer,
-		NewRunnerProvider(command.NewRunner()),
+		NewRunnerProvider(command.NewRunner(), wasm.UnimplementedRuntime),
 	)
 	require.NoError(t, err)
 	duplicateBuiltInRulePluginConfig, err := bufconfig.NewLocalPluginConfig(
@@ -261,6 +264,7 @@ func TestMultiClientCannotHaveOverlappingCategories(t *testing.T) {
 	require.NoError(t, err)
 	multiClient := newMultiClient(
 		zap.NewNop(),
+		tracing.NopTracer,
 		[]*checkClientSpec{
 			newCheckClientSpec("buf-plugin-1", client1, emptyOptions),
 			newCheckClientSpec("buf-plugin-2", client2, emptyOptions),
@@ -279,7 +283,7 @@ func TestMultiClientCannotHaveOverlappingCategoriesWithBuiltIn(t *testing.T) {
 	client, err := newClient(
 		zaptest.NewLogger(t),
 		tracing.NopTracer,
-		NewRunnerProvider(command.NewRunner()),
+		NewRunnerProvider(command.NewRunner(), wasm.UnimplementedRuntime),
 	)
 	require.NoError(t, err)
 	duplicateBuiltInRulePluginConfig, err := bufconfig.NewLocalPluginConfig(